+++ R/finbif_check_taxa.R
@@ -7,7 +7,9 @@
 #' @param taxa Character (or list of named character) vector(s). If a list each
 #'   vector can have the name of a taxonomic rank (genus, species, etc.,).
 #'   The elements of the vectors should be the taxa to check.
-#' @param cache Logical. Use cached data.
+#' @param cache Logical or Integer. If `TRUE` or a number greater than zero,
+#'   then data-caching will be used. If not logical then cache will be
+#'   invalidated after the number of hours indicated by the argument.
 #' @return An object of class `finbif_taxa`. A list with the same form as
 #'   `taxa`.
 #' @examples \dontrun{