+++ R/finbif_taxa.R
@@ -10,7 +10,9 @@
 #'   "likely" only one taxon will be returned.
 #' @param type Character. Type of match to make. Must be one of `exact`,
 #'   `partial` or `likely`.
-#' @param cache Logical. Use cached data.
+#' @param cache Logical or Integer. If `TRUE` or a number greater than zero,
+#'   then data-caching will be used. If not logical then cache will be
+#'   invalidated after the number of hours indicated by the argument.
 #' @param locale Character. One of the supported two-letter ISO 639-1 language
 #'   codes. Current supported languages are English, Finnish and Swedish. For
 #'   data where more than one language is available the language denoted by